chore: centralize fixtures on conftest.py
