+++ tests/transformers/test_tokenizers.py
@@ -10,8 +10,6 @@ from nlpiper.core.document import (
     Token
 )
 
-from tests.transformers import hide_available_pkg  # noqa: F401
-
 
 class TestTokenizersValidations:
